writing-app/requirements.txt
streamlit
easyocr
pillow
pytesseract
transformers
sentencepiece